Prepare for downstream usage
