+++ docs/rst/api.rst
@@ -1,0 +1,12 @@
+API library documentation
+=========================
+
+.. toctree::
+   :maxdepth: 1
+
+   api/cli
+   api/libcli
+   api/save
+   api/colors
+   api/logger
+   api/utils
