+++ docs/rst/api/cli.rst
@@ -1,0 +1,6 @@
+
+High-Level API
+==============
+
+.. automodule:: mdacli.cli
+   :members:
