+++ docs/rst/api/colors.rst
@@ -1,0 +1,5 @@
+Coloring
+========
+
+.. automodule:: mdacli.colors
+   :members:
