+++ docs/rst/api/libcli.rst
@@ -1,0 +1,5 @@
+Support functions for CLI
+=========================
+
+.. automodule:: mdacli.libcli
+   :members:
