+++ docs/rst/api/logger.rst
@@ -1,0 +1,5 @@
+Logging
+=======
+
+.. automodule:: mdacli.logger
+   :members:
