+++ docs/rst/api/save.rst
@@ -1,0 +1,5 @@
+Results Saving
+==============
+
+.. automodule:: mdacli.save
+   :members:
