+++ docs/rst/api/utils.rst
@@ -1,0 +1,5 @@
+Utilities
+=========
+
+.. automodule:: mdacli.utils
+   :members:
