--- docs/rst/reference/index.rst
@@ -1,7 +1,0 @@
-Source documentation
-====================
-
-.. toctree::
-    :glob:
-
-    mdacli*
