--- docs/rst/reference/mdacli.rst
@@ -1,10 +1,0 @@
-mdacli
-======
-
-.. testsetup::
-
-    from mdacli import *
-
-.. automodule:: mdacli
-    :members:
-
