+++ src/mdacli/__main__.py
@@ -7,15 +7,42 @@
 # SPDX-License-Identifier: GPL-2.0-or-later
 
 """
-Entrypoint module, in case you use `python3 -mdacli`.
+A command line interface (CLI) to the analysis modules of MDAnalysis.
 
-Why does this file exist, and why __main__? For more info, read:
+The modules are all structured as part of a single mdacli wrapper, and invoked
+with commands like `mda RMSD`. This command uses the class
+:class:`MDAnalysis.analysis.rms.RMSD` for calculating the RMSD.
+Documentation for each module can be found at the respective sections on the
+`MDAnalysis Analysis Documentation`_, as well as
+`mda command -h`.
 
-- https://www.python.org/dev/peps/pep-0338/
-- https://docs.python.org/3/using/cmdline.html#cmdoption-m
+.. _`MDAnalysis Documentation`:
+   https://docs.mdanalysis.org/stable/documentation_pages/analysis_modules.html
 """
 
-from mdacli.cli import main
+from MDAnalysis.analysis import __all__
+
+import mdacli
+
+
+def main():
+    """Execute main CLI entry point."""
+    # Thse module are currently not supported. Either due a different
+    # structure or due parameters that aree not supported by our parser.
+    skip_mods = ['AnalysisFromFunction',
+                 'HydrogenBondAnalysis',
+                 'WaterBridgeAnalysis',
+                 'Contacts',
+                 'Dihedral',
+                 'PersistenceLength',
+                 'InterRDF_s']
+
+    mdacli.cli(name="MDAnalysis",
+               module_list=__all__,
+               version=mdacli.__version__,
+               description=__doc__,
+               skip_modules=skip_mods,
+               ignore_warnings=True)
 
 
 if __name__ == '__main__':
